ONT-1851 Change `age` int->long

+++ src/main/java/at/rtr/rmbt/request/TestSettingsRequest.java
@@ -151,7 +151,7 @@ public class TestSettingsRequest {
         private Long time;
 
         @JsonProperty("age")
-        private Integer age;
+        private Long age;
 
         @JsonProperty("accuracy")
         private Float accuracy;
